Move mount modules to linux/bsd modules
Also remove cfg-if dependency
