+++ Cargo.toml
@@ -9,7 +9,6 @@ libc = "0.2.29"
 syslog = "3.3.0"
 lazy_static = "0.2.1"
 hostname = "^0.1"
-cfg-if = "^0.1"
 argparse = "*"
 
 [dependencies.prometheus]
