+++ src/get_mounts/mod.rs
@@ -1,9 +1,9 @@
-cfg_if! {
-        if #[cfg(target_os = "linux")] {
-                mod getmntent;
-                pub use self::getmntent::get_mount_points;
-        } else {
-                mod getmntinfo;
-                pub use self::getmntinfo::get_mount_points;
-        }
-}
+#[cfg(target_os = "linux")]
+mod linux;
+#[cfg(target_os = "linux")]
+pub use self::linux::get_mount_points;
+
+#[cfg(all(unix, not(target_os = "linux")))]
+mod bsd;
+#[cfg(all(unix, not(target_os = "linux")))]
+pub use self::bsd::get_mount_points;
